+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * **Add nuxt-buefy on docs**, [see here](https://buefy.github.io/#/documentation/start) (thanks @anteriovieira)
+* Add ``indefinite`` prop to snackbar
 * #509 Add autocomplete slot support to taginput (thanks @gkinsman)
 * #581 Add ``is-row-checkable`` prop to table (thanks @jvmonjo)
 * #602 Add ``changeMonth`` and ``changeYear`` events to datepicker